Separated out the capacity checking part of the capacity report maintainer. Made aws zones excluded from the capacity calculation because the metric is nonsensical for an automatically scaling zone. Fixed some missing renames.
